Assert each test is called only once
Added a counter in each service to start at zero upon initialization and increment each time the service's functionality is executed. Need to implement a metadata for the test service in the event it repeats a call to A; if it is a repeat call to A, the counters should be at 0 instead of 1.